+++ src/data/utils/exportMsg.js
@@ -7,7 +7,7 @@ import isError from '../type/isError'
  * @param {string} [option.data] 额外信息内容
  * @param {'error' | 'warn' | 'log'} [option.type] 额外信息提示类型
  */
-function printMsgAct(msg, type = 'error', option = {}) {
+function exportMsg(msg, type = 'error', option = {}) {
   if (type == 'error') {
     if (!isError(msg)) {
       console[type](new Error(msg))
@@ -25,4 +25,4 @@ function printMsgAct(msg, type = 'error', option = {}) {
   }
 }
 
-export default printMsgAct
+export default exportMsg